split plotters script into general plotters and plots for the pathways ms
